requirements.txt
argparse
jinja2
pdfkit
python-nmap

## wkhtmltopdf